Fix byte utils and value test
Conflicts:

	common/src/test/java/org/openecard/client/common/util/ByteUtilsTest.java
	common/src/test/java/org/openecard/client/common/util/ValueGeneratorsTest.java

+++ common/src/main/java/org/openecard/client/common/util/ByteUtils.java
@@ -298,7 +298,7 @@ public class ByteUtils {
 	long value = 0;
 
 	for (int i = 0; i < bytes.length; i++) {
-	    value |= (0xFF & bytes[bytes.length - 1 - i]) << i * 8;
+	    value |= ((long) 0xFF & bytes[bytes.length - 1 - i]) << i * 8;
 	}
 
 	return value;
